Add old deliverable files
Add deliverable files for existing releases, including the script used
to generate them from launchpad and git data. This data is known to be
incomplete, and is meant as a starting point for reconstructing the full
history.

The output from running the script [1] shows that many projects we
expected to treat as deliverables do not yet have a launchpad page.

[1] http://paste.openstack.org/show/336176

Change-Id: I1a64c8b6b99ec575c00c1ebad93a5f6dc70c9718