validate the current release series by default
If we submit changes to the validation script we want them tested
against something, so default to looking at all of the files for the
current release. Running it against all files for historical releases
would take a long time and might fail if we introduce schema changes to
the deliverables files.

Change-Id: If9b840403f52abde597e1d99b594be338c42aa3d