Refactor validation script
Move some of the validation code into a real python package so we can
more easily reuse it.

Make the validation script reprint all errors, for cases where it has
checked a lot of files and the original messages might have been mixed
in with other output.

Fix the launchpad project name for tooz so the validation script will
pass.

Change-Id: I17534d460dd9ed19d10e48a0ef5d28dac3ab0fe7