Adds oslo.service 0.7.0 deliverable
The 0.6.0 version of the oslo.service causes services using the
library to stop working on Windows. This affects the nova-compute
service on Windows.

The problem is that it tries access signal.SIGHUP which does not
exist on the Windows platform. This issue was fixed in the commit
with the Change-Id: Id3783160381f4a89e241134ef13c4521a30dfa94 but
it is not released.

Change-Id: Ie6bba338ab065aac75323c13cd1dc01b925bd44d