Make list-changes smarter
If we can't figure out the previous version from looking in the YAML
file, use git describe to find it.

Show the version number being released in the header of the list-changes
output so we can verify that it matches the expected value after seeing
what sorts of changes are included.

Change-Id: I8f65400900ee0779fd8d1736bf319e638ebbeca0