Bump python-swiftclient to 2.3.2 for stable/juno
Commit c213ffda17d09a68fdbb619ba6db4b8a101e85a9 synced
python-swiftclient with global-requirements from stable/juno to cap
futures so it doesn't conflict with taskflow (which has futures capped
on in juno per g-r).

Requirements commit f7c3edbc9fa40e6a3979a2dc58c2a9a6af374607 raised the
cap on python-swiftclient in global-requirements to <2.4.0 so this 2.3.2
release will fix the futures dependency conflict.

This change includes the hashes for 2.2.0->2.3.1 since those are the
other possible versions in the supported range for g-r on stable/juno.

Closes-Bug: #1486576

Change-Id: Ib7c987d2eaf46dc18b166d773b8125b834ee5c70