Release python-novaclient 2.28.0
This is mostly needed for some key fixes once we deployed v2.0 on v2.1
with 8349aff5a and 4a18f7d3b which we should release before we close out
liberty.

The version is 2.28.0 rather than 2.27.1 because of 9b7dd38 which looks
like a feature that slipped in without a blueprint or bug.

Related-Bug: #1491579
Related-Bug: #1491325

Change-Id: I978e82a699b9f33898926c13332bb6461a0e4dae