Release python-novaclient 2.28.1
This is to get the fix for 6fc3a94 into a release and then make 2.28.1
the minimum required version for liberty in global-requirements to also
pick up the other fixes for the recent v2.0 on v2.1 failures.

There is another change d970de4 in here but it actually doesn't do
anything since novaclient isn't translated.

Related-Bug: #1492242

Change-Id: Ia633181c806bd0200a869e6b2102e4a9d560c56a