Import the release status table from wiki
Import the name and status column from
https://wiki.openstack.org/wiki/Releases

This makes the toctree redundant, so hide it.

Make the toctree an explicit list of version names, so we can control
the order they appear. This lets us list the most recent version first
and the navigation for the "next" link will take us back in time.

I left out the individual tagged releases and dates because the versions
are available on the release page.

Change-Id: I7a9658cbbb4f6bd854f2938857f8c8ea6e1d5bd6