Add review notes to the readme
We need to document general guidelines for reviewing release requests.

This is following the recent gate breakage with oslo.utils 1.4.1.

The intent here is mostly for the special cases which impact stable branches,
but the semver reminder is important for non-stable branches also since
improper versioning on master where semver wasn't followed, e.g. incrementing
a patch version when the minor version should have been updated, has led to
a lot of the problems we see with stable branches.

Related-Bug: #1496650

Change-Id: I4c7013e265bb533dbac2c8804dea01dacff5de4b