mitaka: release python-novaclient 2.34.0
Where to start...

6bda37f498f5ecb262dbb5d27c3f1ff8a47369d8 requested 2.33.0 which we
needed to get 28bf8dd released.

The 2.33.0 tag is on a5363ae0d60d41feff23f6cca44fcd7aa9fc5b8b so we
wouldn't pick up 0cd5812 which removes the novaclient.v1_1 module and is
a backward incompatible change.

However, 2.33.0 contained 2961e82 which was also a backward incompatible
change, which was reverted with 63c7a57.

2.33.0 has been blacklisted from global-requirements:

I6e7657b60308b30eed89b269810c1f37cce43063

Now we want to release a new version to include that revert. Note that
we reverted the removal of the novaclient.v1_1 module since that would
have also been a backward incompatible change. We reverted that so we
could release this as 2.34 with bug fixes. Once this is released we'll
redo the removal of the novaclient.v1_1 module and release that as
3.0.0.

Changes since 2.33.0:

mriedem@ubuntu:~/git/python-novaclient$ git log --oneline --no-merges
2.33.0..
036d42d Revert "Remove novaclient.v1_1 module"
63c7a57 Revert "Do not expose exceptions from requests library"
217e7c1 Updated from global requirements
0cd5812 Remove novaclient.v1_1 module

Related-Bug: #1510790

Change-Id: I4fc044e518a5cc5ea7b309a2824984a7035945bd