Add section for deliverables with no type
Deliverables which do not have any type were assumed to be
type:service deliverables and were listed with services,
resulting in a crowded and confusing output.

This patch makes sure they are listed in their own section.
It is still wrong for some of them (which should really have the
type:service tag), but less wrong than considering them all as
services.

Change-Id: I828b0fb245443cf5e3fb50c222da3f70c65a014b