Add a warning for new mitaka releases of oslo.messaging
Neutron (and possibly others) were/are using a private symbol of
oslo.messaging  several reviews have been put in place to mitigate this
including a requirements change that bans
!=2.6.0,!=2.6.1,!=2.7.0,!=2.8.0,!=2.8.1,!=2.9.0 from liberty

If a new release is tagged in the range >2.6.0 <3.0.0 it will NOT be
banned and may introduce a regression.

A release in that range is doable but we need to be careful.  Add a
warning to that we don't forget and do that thing.

Depends-On: I5d15366a212c6ed21f33bfe011876fc2bb56997c
Change-Id: Ia7cdbbed5e529864e751add6af05e5fc0be963e1