add list-constraints command
list-constraints takes a series name as input and produces a constraints
file including all of the packages released as part of that series. It
doesn't differentiate between servers and libraries and tools, since we
don't know from here which might be listed in a given constraints
file. Merging the output with a real constraints file is the
responsibility of the user.

Change-Id: I6f669dca320825f7a313953f8d13f5daf5e51368