update validator to handle new repos being added to deliverables
When a new repository is added to a deliverable in the middle of a
series, the validator gets confused trying to determine whether the new
tag is valid compared to the old one, because the old one may not exist
in the repository. Skip that step to allow new projects to be added
mid-cycle.

Change-Id: I05efc1007f906c5d341f802b14ebaae12cd52043
Signed-off-by: Doug Hellmann <doug@doughellmann.com>