rearrange content in preparation to move to new vhost
We want to remove the redundant 'releases' directory in the URLs for
release pages, so create one directory for each series and move the
existing file there. Sphinx doesn't have an easy way to handle
redirects, so just add a stub page with a link to the root document for
folks to find the thing that moved. We will delete these pages after we
start publishing to releases.openstack.org instead of
docs.openstack.org.

Change-Id: I970bdd20e23c748a7e8a456808c73d74570eb066
Signed-off-by: Doug Hellmann <doug@doughellmann.com>