show more change detail, including ancestry
Remove the skip for showing changes for existing tags. We still only
process the final version for a given deliverable file.

Show the git log commands being run at each step.

Use topological ordering for log commands so we get the commits that
will actually be included in a version.

Add ancestry check and report the difference in ancestry. The validation
command should fail if the new tag is not an ancestor, but that check
only runs if the tag has not already been applied to the repository.

Change-Id: I51d0ae570a41ef39d783d7db18b250a4ab7cf954
Signed-off-by: Doug Hellmann <doug@doughellmann.com>