import a copy of openstackdocstheme
This is a forked copy of openstackdocstheme, which will serve as the
basis for the theme for releases.openstack.org. Rather than try to make
the existing theme work, or "subclass" it, I plan to modify it here in
place. That will avoid issues from future changes upstream by the docs
team, who have their own requirements. I'm checking the whole thing in
in this patch, and will modify it in a follow-up patch, to make clear
what the modifications are.

Change-Id: I3cd9c0e44404447b7c6497f8f9f054a94ec1a46f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>