Update python-troveclient deliverable for mitaka
Version 2.1.0 of the python-troveclient included a change that broke
compatibility. This change is to request that 2.1.1 (which has the fix
for the compatibility) be released with Mitaka.

This however is subject to the fact that Trove has a change that was
granted an FFE, and that change has a client side component. That
change (https://review.openstack.org/#/c/290177/) will land and then I
will request 2.2.0 be tagged.

Change-Id: I5f113c7e4a3af50b34c1cceb542ffc5a71d9395c
Related-Bug: #1538506