Change openstack-ansible from independent to cycle-with-intermediary
openstack-ansible needs to be able to release version-specific updates
after a release cycle is finished. Our tools don't support that well,
and they want to change to the cycle-with-intermediary release model
anyway, so let's just undo the change to merge their release data into
an independent deliverable file by restoring the liberty deliverable
file. The project was not official during kilo, so do not restore
that file.

The governance change for the release model update is in
https://review.openstack.org/#/c/297297

This partially reverts commit 5e15c8b6ab49486e242e23d3d028fbf7a85b9360.

Change-Id: I0a9baae82da06841904300894e47cfd05759d0ca
Signed-off-by: Doug Hellmann <doug@doughellmann.com>