fix the way we list changes that will not be included
The git log range is apparently inclusive, which means when we try to
produce the list of changes that won't be in a release, we need to stop
at the parent of the target commit. Use ~1 to indicate the first parent
of the commit, which should always be the commit on the branch into
which a merge happened.

Change-Id: I7189b296a74445779c2227d63809d7d905cbb2db
Signed-off-by: Doug Hellmann <doug@doughellmann.com>