Please tag python-troveclient v2.2.0
This change was earlier proposed in mitaka but it was not timely and
therefore did not make it to Mitaka. I'm resubmitting this for Newton
and will request the bump of g-r once this merges.

This change adds client side support for the trove 'module' facility
that was released in the trove server in Mitaka.

Change-Id: I89f8af415160213edbb3f7587701644dd0342698