invoke sphinx directly instead of through pbr
The pbr feature for invoking sphinx doesn't let us treat warnings as
errors because of a bug in pbr. Switch to invoking sphinx directly,
including removing the setup.cfg instructions for sphinx, so we can use
the flag to treat warnings as errors. Doing this also requires fixing a
few existing warnings, so those changes are included in this patch.

Change-Id: I2aa0d844855d5a18646d0bc3907620544ae71be2
Signed-off-by: Doug Hellmann <doug@doughellmann.com>