relax ancestry rules for independent projects
For cycle-based projects we want to require new releases to be on the
same branch as previous releases in the same series. We do that by
enforcing an ancestry between the commit being tagged and the previous
commits in the series.

For independent projects, we only have one deliverable file for each and
so we need to be able to add new releases from master to the same file
that we're releasing older stable branches. Treat situations where
independent projects have follow-up releases from different branches as
warnings instead of errors.

Fix the existing historical record for hacking by removing the comments.

Change-Id: Ia80c876c8de94259a1a56b9c3d07dbefc3e6d1e5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>