Add a deliverable for puppet-tripleo
Add a deliverable and an initial release (we've never yet tagged one)
for puppet-tripleo.

Note this is going under _independent for now, but we intend on doing
coordinated releases for all tripleo components for each milestone
during the newton cycle ref https://review.openstack.org/#/c/308236/
this is just an initial release that's required for downstream packages
to consume, future releases will be actual newton deliverables.

The hash chosen is based on the most recent current-tripleo delorean,
which is based on our most recent periodic CI job promotion.

Change-Id: I44b0ab5571ac3761c92c49f6ad0ebbb73dec2f71