sort version numbers as we display them
We have been relying on folks adding entries to the deliverable files in
release order. That breaks the scripts that actually prepare the
releases, since those scripts only look at the last entry in each file
and releasing a stable version of an independent project often means
inserting an entry into the middle of the deliverable file.

This change tries to sort the versions in the deliverable files as it
reads them before building any tables. We use a tuple of the parts of
the version string, converted to integers where possible to avoid weird
alphabetical sorting of numbers and because SemanticVersion can't always
parse some of the legacy version tags project teams used.

The patch includes one change to an existing deliverable file to
demonstrate that the output is sorted properly.

Change-Id: I962a38db6ca79be02a3fdc8d387e30ee86a6a61c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>