treat independent projects in the wrong directory as warning
The release model tags for projects change over time. Our current data
model doesn't support that, and assumes that if an old deliverable file
is updated the deliverable will still match the current tags. Until we
have a better way to fix that, treat the case where a file appears to be
in the wrong place as a warning instead of an error.

Change-Id: I9d328e3e40b9dfd7f7e6aea8268d547f937ba9b3
Signed-off-by: Doug Hellmann <doug@doughellmann.com>