Add Support for validating 4 digit verion strings
The xstatic packages have a 4 digit verion number 3-digit semver as
supplied by upstream and a local build number.  PBR only supports 3
digit semver.

To work with that create the concpet of a release type that can have
slightly different validation rules.  In this case use the 'packaging'
library.

This will allow the horizon team to use the standard release process.

For refernce see:
https://blueprints.launchpad.net/horizon/+spec/xstatic-release-process

Change-Id: Ie0f33097c31ee4006ec58147b35731913f7b6f4b