Add support to the history import tools for release-type
When trying to use the history import tools we need to use the correct
release-type setting to validate tags.

We can't use versionutils.validate_version() as it's a generator which is
hard to use in a sorted() function.  Insetad introduce
versionutils.canonical_version() which will return the version is valid
or raise a ValueError() exception.

Change-Id: If10aab4c573342227e2909fd84f843c96f3abc43