Release glance-store 0.15.0
The 0.14.0 failed the glance gate for docs and was blocking the review
queue. So, we need this quick follow up release to fix the glance gate
and use all the commits and fixes proposed therein from the prior one.

More explanation about the same is given in bug 1606746 and bug 1605648

Change-Id: I4d17375d2a716581633a5eb91bd6c44ff0a64ab8