Release glance-store 0.16.0
We've removed the s3 driver from tree and this release is special to
that purpose. It would be nice to see this released and tested in gate
and see if we need further changes in glance_store or glance. So, with
the hope of having a few weeks before the newton freeze on non-client
libraries, we wish to see this released and tested.

Change-Id: Ie2988630e12bd330d12982dccc3297721afb6c73