do not validate ancestry of tags for independent releases
Use the placement of the deliverable file in the _independent directory
as an indication that the deliverable was independent, at least for the
version ranges covered in that file.

Only look at the ancestry of a tag if the deliverable is not
independent, because in that case mixes of series will be included in
one file.

Change-Id: Ib12ec6c15d4fea811c87e77ae4356868a10961fc
Signed-off-by: Doug Hellmann <doug@doughellmann.com>