proposed ocata schedule
This schedule is based on the final Ocata release happening the same
week as the PTG to kick off the Pike cycle, currently scheduled for late
February 2017 about 16 weeks after the design summit in Barcelona. It
gives us 3 milestones with 3/4/6 weeks respectively. It leaves only 1
week in the period after the feature freeze at the 3rd milestone and
before RC-1. This will make it especially important for projects to
minimize the feature freeze exceptions granted.  This schedule does
maintain the usual 2 weeks between RC-1 and the final release.

For comparison, the Liberty cycle was 5/5/5 [1] and Juno was 4/6/6
[2]. There is an option that allows 5/5/5 for Ocata, but using the
proposed schedule makes allowances for the holidays at the end of
December and beginning of January, which wasn't an issue for Liberty
since it happened in the other half of the year.

There is an alternative that would allow 3/4/7 with the release happening
2 weeks after the PTG, but that schedule was rejected by some teams during
the discussion at the Austin summit, in part because more significant
release finalization work would have to be happening at the PTG event.

There is another alternative that would give us 2/5/6 by moving O-1 one
week earlier, but because the TC is asking teams to prepare plans to
handle community-wide goals I thought the extra week before the first
milestone would be good to have.

Dates for election-related events will be added later when the release
and PTG dates are confirmed.

[1] https://wiki.openstack.org/wiki/Liberty_Release_Schedule
[2] https://wiki.openstack.org/wiki/Juno_Release_Schedule

Co-Authored-By: Thierry Carrez <thierry@openstack.org>
Change-Id: I6109e43571f397a190a9055996a462176bd9ebe2
Signed-off-by: Doug Hellmann <doug@doughellmann.com>