Add include-pypi-link to python deliverables
Recent releases of python-magnumclient and python-senlinclient did not
generate the expected 'new-release' reviews in openstack/requirements.
This is because they're missing the include-pypi-link key.

While correcting this I did a quick check of
deliverables/newton/python-*yaml and added it to files that are missing
it *and* are mentioned in upper-constraints.txt

Change-Id: I3d99f9a3891ca14ba3d6591a96d4f4b44f73ef54