Release os-client-config 1.21.1
We have blocks in a bunch of requirements preventing 1.18.0 - 1.21.0 from
being used, but other things in the broader ecosystem, such as ansible,
require 1.20.0. The blocks are there to work around a neutronclient
issue, so once neutronclient is released, we need to cut a new occ so
that we don't have conflicting hard requirements floating around out
there.

Change-Id: I9141c957881751a62b02e57b727524abdedc4efe
Depends-On: I9b64fa6ec4cf4e831d4d64b2e666bebf7a1f00be