Release xstatic-angular-smart-table 1.4.13.2
As part of the improvements to the xstatic management and release
process, we have been using an xstatic-release[1] script to help with
creation and sanitisation of these packages. Unfortunately, an oversight
meant that the metadata of these packages was invalid, causing
downstream packagers to be unable to use them. This release updates the
metadata without changing the content of the package.

Relevant IRC log:
http://eavesdrop.openstack.org/irclogs/%23openstack-release/%23openstack-release.2016-09-06.log.html#t2016-09-06T14:08:03

1. https://pypi.python.org/pypi/xstatic-release/1.2.1

Change-Id: I1433b5c984b0489351d04ac88e244f35ebb42c14