track the branch we're testing separately from the series we're given
Keep track of which branch we're examining when looking at the value of
HEAD so we can print a more sensible message when the SHA for a new tag
matches. This is especially useful for independent projects, so we get
messages like "Request releases from HEAD on stable/foo" instead of
"Request releases from HEAD on stable/_independent".

Change-Id: I4990d4bf2e9cc996e217cac71f56f6ef64cf974b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>