mitaka: release nova 13.1.2
This is a bug fix patch release for nova in the mitaka series.

Change 8825efa has a DocImpact tag in it which probably
didn't need to be in the commit message since the migration
states are not part of the external REST API.

Change 9a97047 has a release note but it's just advertising the
bug that was fixed.

There is a single dependency change but it's just blacklisting
a version of a test dependency.

So while there are a bunch of changes, they are all just bug
fixes so this qualifies as a patch release version bump.

Change-Id: Ie20fa270951b69b700222e54bad7b38509a77d0d