move Ocata final release to 22 Feb 2017
Based on feedback during the fishbowl session held at the Ocata summit
in Barcelona, downstream consumers of the final release would like the
release to happen earlier in the week to give them time to create
packages before the weekend. This change moves the final release date
forward one day from Thursday 23 Feb to Wednesday 22 Feb to accommodate
this request. Impact on project teams is expected to be minimal, because
the release team creates the final tag based on existing release
candidates.

Change-Id: I91834e6b89827f928588932d3ea74f938f5d32a4
Signed-off-by: Doug Hellmann <doug@doughellmann.com>