Release OpenStack-Ansible Liberty/12.2.7
To remind reviewers, The Liberty and Mitaka releases
for OpenStack-Ansible are both tagged 'independent'
and not cycle releases. The same file must therefore
be changed to facilitate the release.

This patch is on top of the Mitaka release patch and
will therefore fail the CI verification of tags until
the Mitaka release patch has merged.

Change-Id: I1a6b2f5d14fbfac2b926d3207f68b60feceec5a7