move release model into deliverable files
Because release models change from series to series, we want to track
the data in the releases repository instead of in the governance
repository. This change adds a new 'release-model' field to the
deliverable files to hold the new value.

The valid release models are documented as part of a new reference
section for the published documentation for the site.

The validator looks only at the deliverable file for the release model,
and no longer looks at the projects.yaml file in the governance
repository. For files in the _independent directory, the release model
is inferred from the filename to avoid having to set it redundantly.

Change-Id: I81198a86da98436dc2f65806aafc5f241340c57f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>