only generate new release information, if there is a new release
When running new_release, I've found that it'll generate a new entry
with a new release number and the old sha if there are no intervening
changes. This proposed code will change that behavior and not generate
a new project deliverable file if there is no difference from the
previous one.

I find this useful when I run the new release command for multiple
projects and branches in that if there are no changes to tag, it
generates no output file.

I'm not positive that this change is all good; for example, I don't
know for sure that there are not situations where one would want to
tag a new release even if there are no changes.

Change-Id: I351ada42dde757f29094462b10011ec4f375d465