add a command to build the release dashboard
This new command will replace the one in the release-tools
repository. It uses the deliverable data for a series to produce a
dashboard for tracking the cycle-with-milestone releases so we can set
up a spreadsheet to collaborate on keeping track of those releases.

The old command included deliverables using other release models, but
those are less important/useful to track in this manner.

Change-Id: Ic613d65ae3d143cc910fec201205f1f84fde761c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>