only validate modified deliverable files
Early on when we only had a few deliverable files it was safe to apply
validation changes to all of the current release series. Now that is no
longer the case, because doing so takes far too much time and the jobs
fail regularly when there are so many repos to check.

We have unit tests for the validation rules now, so it is safer to make
changes. This commit changes the default behavior so that if the commit
under review has no deliverable files changed the validate job does
nothing.

Change-Id: I3ab4b71fb245661579cc528526b42f592ef32f7f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>