add more detail to the ics output files
Put the titles of the cross-project events in the summary of the
calendar event so they show up in the main calendar view.

Add the full descriptions of the cross-project events to the calendar
event description so any extra notes (such as specific dates for
deadlines) or details about the event are visible to calendar users.

Change-Id: I6eebcdd827241f008522065a40b765b2089e2856
Signed-off-by: Doug Hellmann <doug@doughellmann.com>