be smarter about required branch for current series
In the period between the time when we start creating stable branches
and we update the default series setting it is possible to release from
the wrong branch. Update the check_branch_sha() logic to look for an
existing stable branch and use that, and only default to master if the
series matches and there is no stable branch yet.

Change-Id: I80b0f39c21b55fb8037c1309392f9deaecef20e7
Signed-off-by: Doug Hellmann <doug@doughellmann.com>