compare tag location with head of its branch
Instead of assuming the master branch, use the actual branch containing
the tag when looking for unreleased changes. This makes the list-changes
report more useful release candidates on a stable branch during the
period when the default series has not been changed (so releases on
stable/a show the right info when the default branch is not b).

Change-Id: I0b5857cce27cd021c59bafe26abfb471f1899311
Signed-off-by: Doug Hellmann <doug@doughellmann.com>