reduce meaningless changes when updating release notes links
If the existing link matches the expected link, leave it in place
regardless of where it is in the file.

If the link needs to be updated, try to insert the link after the
"team" entry, assuming that will be closer to the top of the file
than "releases" and avoiding the need to figure out if "branches" comes
before or after "releases".

Change-Id: I4161bb048deba91ccf8170ea5f0b7208bd9a80a3
Signed-off-by: Doug Hellmann <doug@doughellmann.com>