Move tripleo-ui to cycle-trailing
The TripleO UI is a new deliverable from Newton and should be described
as cycle-trailing, similar to most other TripleO components as they are
released together. Although the project may have cut additional
intermediary releases while getting up and running in Newton, it is no
longer the case as of Ocata.

Having the wrong release model documented is causing some issues. For
example the i18n team assumed there were no cycle-trailing project on
Zanata based on the list of deliverables and planned a week-long outage
[1] during the cycle-trailing RC period, when there may still have been
translation work ongoing.

[1] http://lists.openstack.org/pipermail/openstack-i18n/2017-February/002790.html

Change-Id: I741e8ca63c7555357a7e8b591785a8f6ce12693b