Add Pike deadline details for keystone
Following a very similar release template as the Ocata release,
set project specific dates for keystone. The only thing different
between the proposed schedule for Pike and Ocata is that the
deadline for feature proposal freeze is one week earlier from what
it was in Ocata. This was done to allow for more time between feature
proposal freeze and feature freeze.

Change-Id: Idb2a3b925a4fdf92afef81d1360ffd87a8192d17