Add details of project transitions between stable phases
With Ocata being a short cycle there is some confusion about when
various releases transition from Phase I -> Phase II etc.  This change
adds a column to the main table on releases.o.o to show what the next
phase is and approximately when it will happen.

Mitaka:
    Phase   I:   2016-04-07 Release date
    Phase  II:   2016-10-06 Newton Release date
    EOL      :   2017-04-10 ~1 year after release
Newton:
    Phase   I:   2016-10-06 (Release date)
    Phase  II:   2017-02-22 Ocata Release date[1]
    Phase III:   2017-08-28 ~6months after Phase II
    EOL      :   2017-10-11 ~1 year after release
Ocata:
    Phase   I:   2017-02-22 (Release date)
    Phase  II:   2017-08-28 ~Pike release date[2]
    EOL      :   2018-02-26 ~1 year after release

[1] We only have a single 'Latest Release'
[2] https://releases.openstack.org/pike/schedule.html

Change-Id: Ic48feb9c29273be2fb82ebe723684c56f01d1f67