Gracefully fail if branches have no tags
If some of the branches have no tags at all (like in the case
of feature branches or early master branches), the listing of
branches in list-changes fails.

This problem was introduced in I4f06d6cfaa8ea332d9326ccd8a5e69c749589529
since we switched to subprocess' check_output instead of call.

Change-Id: Id8c61662ecf3a9424c3b4e21fb36f0ed8c01d6b1