Create stable/ocata branch for openstack-manuals
Note that this is the first time openstack-manuals is using this
process. Since 2012, the team only created branches. Since
openstack-manuals now uses reno, let's branch and tag together.

We use 15.0.0 as release number since Ocata is the fifteenth release of
OpenStack.

Note that openstack-manuals has no release-job, we do not push the
content anywhere, just publish.

Change-Id: I0f92f2bf6c563e9ca71fdfba5e00da73d3a98fc7