Release neutron-lib 1.6.0
Change Ic455e1af2796e58025381160dc5c3b83217413fa recently landed in
neutron-lib which rehomes an integral part of our API efforts. We'd like
a new release of lib ASAP so that we can consume the change(s) early-on
in the release cycle to reduce tail-end risk involved with that
transition.

In addition I9966c90e3f90552b41ed84a68b19f3e540426432 is included in
this release and is required to finish up our neutron-lib callback
efforts.

Change-Id: I7f3d4d7451930fcc1d66b331ddc53d24cb86666a