add stable-branch-type option
The rules for stable branches vary based on whether a project is tagged
or not. The default rule set still relies on using tags, but projects
like devstack and grenade will want to be able to create branches
without having to tag. This patch adds a new stable-branch-type option
to let us control the behavior for each deliverable file.

Change-Id: I13c22d9da303cd3a6df329bf310a2292cfec9ae0
Signed-off-by: Doug Hellmann <doug@doughellmann.com>