Request a library release before milestone-2
In order to reduce no-show risk at library freezes, we'd like to make
sure that we have one library release in the cycle to fall back to.
This is done by asking projects to do at least one library release
before milestone-2. Add corresponding steps to the PROCESS.rst file.

Change-Id: I4f64f931bf944ce5d7a9dfdb99dc458c222b8fcb