Switch instack-undercloud to cycle-with-intermediary
This is required for If07880e4bc25cffe485704772d64f62288a1a53d so
we can stop using the bash layer of indirection between tripleoclient
and instack-undercloud.  In order to do that, we need to be able to
depend on instack-undercloud directly, but it can't go into g-r
with its current release model.  This release model better reflects
the intended use of instack-undercloud as a library instead of a
CLI in its own right.

Change-Id: I42235d13acf617f3bcbf22cd737ebf21378902b4