ensure version numbers always have quotes
Update the reformatter to always put quotes around strings that look
like they are numbers. A value like x.y.z is clearly not a float so
the library does the right thing but x.y seems like a float so it
leaves the quotes off, then later when we parse the file again we get
a number not a string.

Change-Id: I6126f59d09d169a2dde1cdffcecd15f8c65c2e65
Signed-off-by: Doug Hellmann <doug@doughellmann.com>