add watched queries to list-changes
Add a data file for defining gerrit queries so we can track incomplete
work for known initiatives when reviewing releases. Include a query
for patches owned by the proposal bot (to watch for requirements
updates and other automated changes) and another to watch for
documentation file updates (to let us monitor the doc migration
initiative for pike).

Change-Id: I12c59d83267bf80773876e81b3470bfe3062ad27
Signed-off-by: Doug Hellmann <doug@doughellmann.com>